Implement response penalty for bot messages in DecideToRespond
Added a response change penalty for messages originating from other bots in the DecideToRespond class. The penalty factor is set to a default value of 0.4, which can be overridden by the BOT_RESPONSE_CHANGE_PENALTY environment variable. Enhanced logging and exception handling are also included to ensure robust operation and easier debugging.